requirements.txt
mnemonic
bip_utils
qrcode
fpdf2
Pillow